+++ packages/plugins/didkit/CHANGELOG.md
@@ -1,5 +1,13 @@
 # learn-card-core
 
+## 1.4.19
+
+### Patch Changes
+
+-   Updated dependencies [[`20d4585c3a2bc8c5eb4b0a628eb215be829000fa`](https://github.com/learningeconomy/LearnCard/commit/20d4585c3a2bc8c5eb4b0a628eb215be829000fa)]:
+    -   @learncard/types@5.5.9
+    -   @learncard/core@9.3.10
+
 ## 1.4.18
 
 ### Patch Changes