+++ packages/plugins/learn-card-network/CHANGELOG.md
@@ -1,5 +1,19 @@
 # learn-card-core
 
+## 2.2.3
+
+### Patch Changes
+
+-   [#580](https://github.com/learningeconomy/LearnCard/pull/580) [`9b4e281b9a07c461fd3b9c556225282817ca3861`](https://github.com/learningeconomy/LearnCard/commit/9b4e281b9a07c461fd3b9c556225282817ca3861) Thanks [@TaylorBeeston](https://github.com/TaylorBeeston)! - Update type/implementation of `updateBoost` to not require a credential when updating
+
+## 2.2.2
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/network-brain-client@2.1.2
+    -   @learncard/core@9.3.10
+
 ## 2.2.1
 
 ### Patch Changes